+++ tests/golden/syscall-hello-elf-o1-c/source.c
@@ -18,7 +18,7 @@
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
+long _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
     int64_t v0 = 0LL;
     int64_t v2 = 0LL;
     int8_t v5 = 0LL;
@@ -38,7 +38,7 @@ int64_t _init(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t ar
         v12 = ((long long (*)(long long, long long, long long, long long, long long, long long))(/* opaque: indirect-call */ 0))(v6, v7, v8, v9, v10, v11);
     }
     /* phi v13 <- (bb0: v2) (bb1: v12) */
-    return v13;
+    return ((long)(v13));
 }
 
 /* dac-recovered function */
@@ -114,7 +114,7 @@ void _start(void) {
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t deregister_tm_clones(void) {
+long deregister_tm_clones(void) {
     int8_t v2 = 0LL;
     int64_t v3 = 0LL;
     int8_t v6 = 0LL;
@@ -124,7 +124,7 @@ int64_t deregister_tm_clones(void) {
     if (v2) {
 L0:;
         /* phi v7 <- (bb0: 16400) (bb1: v3) */
-        return v7;
+        return ((long)(v7));
     } else {
         v3 = (*((int64_t *)(16328LL)));
         v6 = (v3 == 0LL);
@@ -150,7 +150,7 @@ L0:;
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t register_tm_clones(void) {
+long register_tm_clones(void) {
     int64_t v8 = 0LL;
     int8_t v11 = 0LL;
     int64_t v12 = 0LL;
@@ -172,7 +172,7 @@ int64_t register_tm_clones(void) {
 /* stack_locals: 1 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
+long __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) {
     int8_t v0 = 0LL;
     int8_t v1 = 0LL;
     int64_t v2 = 0LL;
@@ -195,7 +195,7 @@ int64_t __do_global_dtors_aux(int64_t arg0, int64_t arg1, int64_t arg2, int64_t 
     v0 = (*((int8_t *)(16400LL)));
     v1 = (v0 != 0LL);
     if (v1) {
-        return v21;
+        return ((long)(v21));
     } else {
         v3 = ((void *)((v2 - 8LL)));
         *((int64_t *)(((int64_t)(v3)))) = v4;
@@ -227,9 +227,9 @@ void frame_dummy(void) {
 /* stack_locals: 0 */
 /* struct_layouts: pointer=0 stack=0 */
 /* switch_tables: 0 */
-int64_t _fini(void) {
+long _fini(void) {
     int64_t v0 = 0LL;
     int64_t v3 = 0LL;
 
-    return v3;
+    return ((long)(v3));
 }
